Extract users to separate module
