+++ tests/routers/openml/datasets_list_datasets_test.py
@@ -9,7 +9,7 @@ from hypothesis import strategies as st
 from starlette.testclient import TestClient
 
 from tests import constants
-from tests.routers.openml.users_test import ApiKey
+from tests.users import ApiKey
 
 
 def _assert_empty_result(
